Move waterfall task status icon
